# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flirt & Play – game platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /**
     * @minLength 2
     * @maxLength 30
     */
  username: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  email: string;
  username: string;
  /** @nullable */
  avatarPath?: string | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  /**
     * @minLength 2
     * @maxLength 30
     */
  username?: string;
  /** @nullable */
  avatarPath?: string | null;
}

export interface UserStats {
  wins: number;
  losses: number;
  gamesPlayed: number;
  flirtsSent: number;
  flirtsReceived: number;
}

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  waiting: 'waiting',
  active: 'active',
  finished: 'finished',
} as const;

export interface Room {
  id: number;
  code: string;
  status: RoomStatus;
  creatorUserId: number;
  /** @nullable */
  guestUserId?: number | null;
  creatorProfile?: UserProfile;
  guestProfile?: UserProfile;
  createdAt: string;
}

export interface GameType {
  key: string;
  name: string;
  description: string;
  minPlayers: number;
  maxPlayers: number;
}

export type GameSessionStatus = typeof GameSessionStatus[keyof typeof GameSessionStatus];


export const GameSessionStatus = {
  active: 'active',
  finished: 'finished',
} as const;

/**
 * Game-specific board/state as JSON
 */
export type GameSessionState = { [key: string]: unknown };

export interface GameSession {
  id: number;
  roomId: number;
  gameType: string;
  playerXUserId: number;
  playerOUserId: number;
  currentTurnUserId: number;
  /** @nullable */
  winnerUserId?: number | null;
  isDraw?: boolean;
  status: GameSessionStatus;
  /** Game-specific board/state as JSON */
  state: GameSessionState;
  createdAt: string;
  /** @nullable */
  finishedAt?: string | null;
  playerXProfile?: UserProfile;
  playerOProfile?: UserProfile;
  flirtSent?: boolean;
}

export type GameInputGameType = typeof GameInputGameType[keyof typeof GameInputGameType];


export const GameInputGameType = {
  'tic-tac-toe': 'tic-tac-toe',
  'connect-four': 'connect-four',
  'rock-paper-scissors': 'rock-paper-scissors',
  'word-duel': 'word-duel',
  'truth-spinner': 'truth-spinner',
} as const;

export interface GameInput {
  gameType: GameInputGameType;
}

export type GameMoveMoveData = { [key: string]: unknown };

export interface GameMove {
  id: number;
  gameId: number;
  playerUserId: number;
  moveData: GameMoveMoveData;
  createdAt: string;
}

export type MoveInputMoveData = { [key: string]: unknown };

export interface MoveInput {
  moveData: MoveInputMoveData;
}

export interface FlirtMessage {
  id: number;
  gameId: number;
  fromUserId: number;
  toUserId: number;
  message: string;
  createdAt: string;
  fromProfile?: UserProfile;
  toProfile?: UserProfile;
}

export interface FlirtMessageInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  message: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type Logout200 = {
  ok?: boolean;
};