# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Flirt & Play – game platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created and signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Validation error or email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /rooms/{roomId}/games:
    get:
      operationId: listRoomGames
      tags: [games]
      summary: Get game history for a room (newest first)
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of game sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameSession"
    post:
      operationId: createGame
      tags: [games]
      summary: Start a new game in the room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Game session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSession"
        "400":
          description: Validation error or room not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /rooms/{roomId}/games/current:
    get:
      operationId: getCurrentGame
      tags: [games]
      summary: Get the active game session for a room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Current active game session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSession"
        "404":
          description: No active game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /games/{gameId}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a game session by ID
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSession"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /games/{gameId}/moves:
    get:
      operationId: listMoves
      tags: [games]
      summary: Get all moves for a game (for playback)
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of moves
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameMove"
    post:
      operationId: makeMove
      tags: [games]
      summary: Make a move in a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveInput"
      responses:
        "200":
          description: Updated game session after move
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSession"
        "400":
          description: Not your turn or invalid move
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden – not a player in this game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /games/{gameId}/rematch:
    post:
      operationId: requestRematch
      tags: [games]
      summary: Request a rematch – creates new game of same type in room
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: New game session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSession"
        "400":
          description: Game not finished yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ── Flirt Messages ────────────────────────────────────────────────────────
  /games/{gameId}/flirt:
    get:
      operationId: getFlirtMessage
      tags: [flirt]
      summary: Get the flirt message sent after this game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Flirt message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlirtMessage"
        "404":
          description: Not sent yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: sendFlirtMessage
      tags: [flirt]
      summary: Send a flirt message from the loser to the winner
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlirtMessageInput"
      responses:
        "201":
          description: Flirt message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlirtMessage"
        "400":
          description: Validation error or game not finished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ── Object Storage ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth ──────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [email, password, username]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        username:
          type: string
          minLength: 2
          maxLength: 30
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    # ── User ──────────────────────────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, email, username, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        avatarPath:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    GameSession:
      type: object
      required: [id, roomId, gameType, playerXUserId, playerOUserId, currentTurnUserId, status, state, createdAt]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        gameType:
          type: string
        playerXUserId:
          type: integer
        playerOUserId:
          type: integer
        currentTurnUserId:
          type: integer
        winnerUserId:
          type: ["integer", "null"]
        isDraw:
          type: boolean
        status:
          type: string
          enum: [active, finished]
        state:
          type: object
          description: Game-specific board/state as JSON
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: ["string", "null"]
          format: date-time
        playerXProfile:
          $ref: "#/components/schemas/UserProfile"
        playerOProfile:
          $ref: "#/components/schemas/UserProfile"
        flirtSent:
          type: boolean
    GameInput:
      type: object
      required: [gameType]
      properties:
        gameType:
          type: string
          enum: [tic-tac-toe, connect-four, rock-paper-scissors, word-duel, truth-spinner]

    # ── Move ──────────────────────────────────────────────────────────────
    GameMove:
      type: object
      required: [id, gameId, playerUserId, moveData, createdAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        playerUserId:
          type: integer
        moveData:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    MoveInput:
      type: object
      required: [moveData]
      properties:
        moveData:
          type: object
          additionalProperties: true

    # ── Flirt ─────────────────────────────────────────────────────────────
    FlirtMessage:
      type: object
      required: [id, gameId, fromUserId, toUserId, message, createdAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: integer
        message:
          type: string
        createdAt:
          type: string
          format: date-time
        fromProfile:
          $ref: "#/components/schemas/UserProfile"
        toProfile:
          $ref: "#/components/schemas/UserProfile"
    FlirtMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 500

    # ── Storage ───────────────────────────────────────────────────────────
...