# lib/api-zod/src/generated/types/flirtMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flirt & Play – game platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from './userProfile';

export interface FlirtMessage {
  id: number;
  gameId: number;
  fromUserId: number;
  toUserId: number;
  message: string;
  createdAt: Date;
  fromProfile?: UserProfile;
  toProfile?: UserProfile;
}