# lib/api-zod/src/generated/types/gameMove.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flirt & Play – game platform API
 * OpenAPI spec version: 0.1.0
 */
import type { GameMoveMoveData } from './gameMoveMoveData';

export interface GameMove {
  id: number;
  gameId: number;
  playerUserId: number;
  moveData: GameMoveMoveData;
  createdAt: Date;
}