# lib/api-zod/src/generated/types/gameSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flirt & Play – game platform API
 * OpenAPI spec version: 0.1.0
 */
import type { GameSessionState } from './gameSessionState';
import type { GameSessionStatus } from './gameSessionStatus';
import type { UserProfile } from './userProfile';

export interface GameSession {
  id: number;
  roomId: number;
  gameType: string;
  playerXUserId: number;
  playerOUserId: number;
  currentTurnUserId: number;
  /** @nullable */
  winnerUserId?: number | null;
  isDraw?: boolean;
  status: GameSessionStatus;
  /** Game-specific board/state as JSON */
  state: GameSessionState;
  createdAt: Date;
  /** @nullable */
  finishedAt?: Date | null;
  playerXProfile?: UserProfile;
  playerOProfile?: UserProfile;
  flirtSent?: boolean;
}