# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flirt & Play – game platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  email: string;
  password: string;
}