# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flirt & Play – game platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /**
     * @minLength 2
     * @maxLength 30
     */
  username: string;
}